Added photo of homebrew keyset connected as gamepad
For use with Augterm Java application, this circuit controls the 4 gamepad buttons and the X axis of the joystick.  3 resistors are used to provide resistance of 100K ohms when when open and 44K ohms (by adding 33K and 47K in parallel with the 100K) when the pinky button is closed.  Note: The code for the KeysetDriver needed to be modified to work with this circuit.